+++ ios/RNAppsFlyer.h
@@ -8,11 +8,19 @@
 #import "RCTEventDispatcher.h"
 #endif
 
+#import "AppsFlyerAttribution.h"
+#import <objc/message.h>
+//#if __has_include(<AppsFlyerLib/AppsFlyerLib.h>) // from Pod
+//#import <AppsFlyerLib/AppsFlyerLib.h>
+//#else
+//#import "AppsFlyerLib.h"
+//#endif
+typedef void (*bypassDidFinishLaunchingWithOption)(id, SEL, NSInteger);
 
 
-@interface RNAppsFlyer : RCTEventEmitter <RCTBridgeModule>
-
+@interface RNAppsFlyer: RCTEventEmitter <RCTBridgeModule, AppsFlyerLibDelegate, AppsFlyerDeepLinkDelegate>
 @end
+
 
 static NSString *const NO_DEVKEY_FOUND              = @"No 'devKey' found or its empty";
 static NSString *const NO_APPID_FOUND               = @"No 'appId' found or its empty";
@@ -21,7 +29,7 @@ static NSString *const EMPTY_OR_CORRUPTED_LIST      = @"No arguments found or li
 static NSString *const SUCCESS                      = @"Success";
 static NSString *const INVALID_URI                  = @"Invalid URI";
 static NSString *const IOS_14_ONLY                  = @"Feature only supported on iOS 14 and above";
-
+    
   // Appsflyer JS objects
   #define afDevKey                                  @"devKey"
   #define afAppId                                   @"appId"
@@ -40,6 +48,8 @@ static NSString *const IOS_14_ONLY                  = @"Feature only supported o
   #define afOnAppOpenAttribution          @"onAppOpenAttribution"
   #define afOnInstallConversionFailure    @"onInstallConversionFailure"
   #define afOnInstallConversionDataLoaded @"onInstallConversionDataLoaded"
+  #define afDeepLink                      @"onDeepLinkListener"
+  #define afOnDeepLinking                 @"onDeepLinking"
 
 // User Invites, Cross Promotion
 #define afCpAppID                       @"crossPromotedAppId"